Created project + Created project browser UI
Created the project. Engine is a C++ static library. Editor is a C# WPF project.
Setup so that when the projects are build, the files are in the same directory, to make for easier linking.

Project browser UI made using WPF. Separate user controls for the open/create project panels